+++ src/vm/vm.cpp
@@ -8,7 +8,7 @@ void vm::vm_init_entry(const std::vector<std::string>& strs,
                        const std::vector<f64>& nums,
                        const std::vector<nasal_builtin_info>& natives,
                        const std::vector<opcode>& code,
-                       const std::unordered_map<std::string, u64>& global_symbol,
+                       const std::unordered_map<std::string, u32>& global_symbol,
                        const std::vector<std::string>& filenames,
                        const std::vector<std::string>& argv) {
     const_number = nums.data();
